move cidr generation to manager
